+++ tests/integration/conftest.py
@@ -15,6 +15,7 @@
 #  along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """Configuration for craft-application integration tests."""
 
+import atexit
 import os
 import pathlib
 import sys
@@ -99,3 +100,41 @@ def pretend_jammy(mocker) -> None:
 @pytest.fixture
 def hello_repository_lp_url() -> str:
     return "https://git.launchpad.net/ubuntu/+source/hello"
+
+
+_registered_exit_funcs = []
+"""Tracks atexit functions registered by craft application."""
+
+
+@pytest.fixture(autouse=True, scope="session")
+def track_atexit_register():
+    """Tracks atexit.register calls from craft application."""
+    original_register = atexit.register
+
+    def tracking_wrapper(func, *args, **kwargs):
+        """Track atexit calls from craft application and pass-through others."""
+        if func.__module__.startswith("craft_application."):
+            # Only track functions from craft_application
+            _registered_exit_funcs.append(func)
+            return None
+        # atexit.register from pytest is passed through
+        return original_register(func, *args, **kwargs)
+
+    # a context manager is required for a session-scoped monkeypatch fixture
+    with pytest.MonkeyPatch.context() as mp:
+        mp.setattr(atexit, "register", tracking_wrapper)
+        yield
+
+
+@pytest.fixture(autouse=True)
+def call_atexit_funcs(monkeypatch):
+    """Calls atexit functions registered by craft application.
+
+    This is needed because atexit functions are not called between parametrized runs of a test.
+    """
+    yield
+    # If CRAFT_DEBUG is set, the state service won't be able to clean up properly.
+    monkeypatch.delenv("CRAFT_DEBUG", raising=False)
+    for func in _registered_exit_funcs:
+        func()
+    _registered_exit_funcs.clear()